Se agregaro tabla de subcategoria se corrigieron relaciones y tabla de categoría.

+++ database/migrations/2020_11_03_165034_create_logs_table.php
@@ -17,7 +17,7 @@ class CreateLogsTable extends Migration
             $table->id();
             $table->unsignedInteger('user_id');
 
-            $table->unsignedSmallInteger('operacion_id');
+            $table->unsignedInteger('operacion_id');
             $table->foreign('operacion_id')->references('id')->on('operaciones');
 
             $table->unsignedBigInteger('registro_id')
